Add files via upload
I mentioned two different product works. In the listing Audit, most of the listing instructions were not followed by the client. In the title, the main keywords are not added. As everybody knows, the brand name should be the first word of the title. The listing is not according to Amazon listing instructions.
In the second PDF, I used my method of work on how to SEO the listing. How to operate with the listing. You can check my method which is very difficult. Most people do not know about the roots, frequency, and relevancy of the keywords. I optimized the listing with logic.